Update macro_v2
The category in the current spreadsheet for unspecified toxin is "Unknown " (with the extra space). Changed category definition in macro from "Unspecified Toxin" to this. 

When we eventually finalize the spreadsheet they should probably both be changed back to "Unspecified Toxin", but in the meantime I think this is the fastest way to avoid miscalculations and confusions (without this fix, Unknown is automatically considered 0 points whereas it should be +2)